+++ packages/ui5-config/src/index.ts
@@ -1,6 +1,7 @@
 export { UI5Config } from './ui5config';
 export {
     Configuration,
+    CustomItem,
     CustomTask,
     CustomMiddleware,
     FioriAppReloadConfig,
@@ -18,5 +19,5 @@ export {
     UI5ProxyConfigTarget
 } from './types';
 export { UI5_DEFAULT } from './defaults';
-export { mergeObjects, getEsmTypesVersion, getTypesVersion, getTypesPackage } from './utils';
+export { mergeObjects, getEsmTypesVersion, getTypesVersion, getTypesPackage, replaceEnvVariables } from './utils';
 export { errorCode as yamlErrorCode, YAMLError } from '@sap-ux/yaml';