+++ packages/cards-editor-config-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/cards-editor-config-writer
 
+## 0.3.3
+
+### Patch Changes
+
+-   Updated dependencies [e3d2324c]
+    -   @sap-ux/ui5-config@0.22.4
+
+## 0.3.2
+
+### Patch Changes
+
+-   Updated dependencies [7f8105c7]
+    -   @sap-ux/ui5-config@0.22.3
+
 ## 0.3.1
 
 ### Patch Changes