+++ packages/fiori-generator-shared/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/fiori-generator-shared
 
+## 0.0.6
+
+### Patch Changes
+
+-   Updated dependencies [e3d2324c]
+    -   @sap-ux/project-access@1.20.0
+
+## 0.0.5
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.14
+
 ## 0.0.4
 
 ### Patch Changes