+++ packages/inquirer-common/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/inquirer-common
 
+## 0.2.8
+
+### Patch Changes
+
+-   Updated dependencies [5d4c0dec]
+    -   @sap-ux/ui5-info@0.5.0
+
+## 0.2.7
+
+### Patch Changes
+
+-   4098bed2: Moves some prompting code to inquirer-common. Adds new module odata-service-inquirer
+
 ## 0.2.6
 
 ### Patch Changes