+++ packages/mockserver-config-writer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/mockserver-config-writer
 
+## 0.5.2
+
+### Patch Changes
+
+-   Updated dependencies [e3d2324c]
+    -   @sap-ux/ui5-config@0.22.4
+
+## 0.5.1
+
+### Patch Changes
+
+-   Updated dependencies [7f8105c7]
+    -   @sap-ux/ui5-config@0.22.3
+
+## 0.5.0
+
+### Minor Changes
+
+-   8db928db: odata-service-writer now uses logic from mockserver-config-writer to generate mockserver
+
 ## 0.4.2
 
 ### Patch Changes