+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/project-access
 
+## 1.20.0
+
+### Minor Changes
+
+-   e3d2324c: Improvements for consumption
+
+### Patch Changes
+
+-   Updated dependencies [e3d2324c]
+    -   @sap-ux/ui5-config@0.22.4
+
+## 1.19.14
+
+### Patch Changes
+
+-   Updated dependencies [7f8105c7]
+    -   @sap-ux/ui5-config@0.22.3
+
 ## 1.19.13
 
 ### Patch Changes