+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.2.14
+
+### Patch Changes
+
+-   Updated dependencies [5d4c0dec]
+    -   @sap-ux/ui5-info@0.5.0
+    -   @sap-ux/inquirer-common@0.2.8
+
+## 0.2.13
+
+### Patch Changes
+
+-   Updated dependencies [4098bed2]
+    -   @sap-ux/inquirer-common@0.2.7
+
 ## 0.2.12
 
 ### Patch Changes