+++ packages/ui5-library-reference-writer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/ui5-library-reference-writer
 
+## 0.0.12
+
+### Patch Changes
+
+-   Updated dependencies [e3d2324c]
+    -   @sap-ux/project-access@1.20.0
+    -   @sap-ux/ui5-config@0.22.4
+
+## 0.0.11
+
+### Patch Changes
+
+-   Updated dependencies [7f8105c7]
+    -   @sap-ux/ui5-config@0.22.3
+    -   @sap-ux/project-access@1.19.14
+
 ## 0.0.10
 
 ### Patch Changes